feat: implement the remaining builtins

+++ crates/revmc-builtins/src/ir.rs
@@ -243,8 +243,8 @@ builtins! {
     DataCopy       = __revmc_builtin_data_copy(@[ecx] ptr, @[sp] ptr) Some(u8),
     ReturnDataLoad = __revmc_builtin_returndataload(@[ecx] ptr, @[sp] ptr) None,
 
-    EofCreate      = __revmc_builtin_eof_create(@[ecx] ptr, @[sp] ptr, u8) Some(u8),
-    ReturnContract = __revmc_builtin_return_contract(@[ecx] ptr, @[sp] ptr, u8) Some(u8),
+    EofCreate      = __revmc_builtin_eof_create(@[ecx] ptr, @[sp] ptr, usize) Some(u8),
+    ReturnContract = __revmc_builtin_return_contract(@[ecx] ptr, @[sp] ptr, usize) Some(u8),
     Create         = __revmc_builtin_create(@[ecx] ptr, @[sp_dyn] ptr, u8, u8) Some(u8),
     Call           = __revmc_builtin_call(@[ecx] ptr, @[sp_dyn] ptr, u8, u8) Some(u8),
     ExtCall        = __revmc_builtin_ext_call(@[ecx] ptr, @[sp_dyn] ptr, u8) Some(u8),
